fix(yap): YV107 — measure the true rate PER INTERVAL, and rebase onto main
Review round 4's BLOCKING: measure_true_rate divided Δcaptured by Δhost
across the ENDPOINTS of the longest monotonic run, so every interval the
device slept through sat in the denominator with none of its audio in the
numerator. A perfect 0 ppm crystal with one 10 s stall in 2 h reported
-1,388.9 ppm and flagged it; a far side silent 5 s in every 15 — this
item's own 'routine shape of the far side' — reported -333,240.7 ppm.
Both reproduced first, both now +0.0 ppm.

measure_run keeps the intervals in which the device was DELIVERING
(TRUE_RATE_INTERVAL_TOLERANCE_SAMPLES, two-sided so a backlog flush is
not a fast crystal) and drops the rest from BOTH sides of the ratio;
TrackRate gains intervals_skipped and ppm_uncertainty, and a track is
never flagged on a number its own index sequence cannot resolve.

Rebased onto main (b2fbb3a). The semantic conflict — this branch's
speaker_label(usize) vs YV108's shipped speaker_label(i64) — is resolved
toward main's, with the_merge_labels_through_the_shipped_render_rule as
the regression guard, and YV109's tests/support/two_track.rs is merged
with this item's rather than forked.

+++ desktop/src-tauri/src/meeting_asr.rs
@@ -1155,10 +1155,22 @@ pub fn merge_two_tracks_by_host_time(
     out
 }
 
+/// The one place a track number becomes a speaker name in the merge.
+///
+/// The cast is the seam between two shipped numberings that are both correct
+/// and are NOT the same type: [`crate::meeting::MIC_TRACK`] is a `usize`
+/// because it indexes the journal's per-track vectors, and
+/// [`crate::meetings::MIC_TRACK`] is an `i64` because it is a SQLite `INTEGER`
+/// column. YV108 shipped `meetings::speaker_label` against the DB numbering and
+/// the UI's TypeScript mirror is held to it; calling it here — rather than
+/// re-deriving "Me"/"Them" against the journal numbering, which would compile
+/// and would be a third copy of the rule — is what keeps the merge, the
+/// Markdown export and the screen from ever disagreeing about who spoke.
+/// `the_merge_labels_through_the_shipped_render_rule` is the guard.
 fn label(track: usize, start_seconds: f64, end_seconds: f64, text: String) -> MergedSpan {
     MergedSpan {
         track,
-        speaker: crate::meetings::speaker_label(track).to_string(),
+        speaker: crate::meetings::speaker_label(track as i64).to_string(),
         start_seconds,
         end_seconds,
         text,
